Output should not corrupt irrespective of documentation content
This happens because the placeholders are declared from a template.
This particular test fails for the struct builder. That's because it replaces the the documentation placeholder first.

```
Template.Replace("{documentation}", SummaryBlock(data.Summary))
```
The same test would pass for the class builder because it replaces the documentation placeholder at the very end.
However, changing the order would be a band-aid fix, because I might have a placeholder as part of a field summary.
A more permanent fix might be to refactor each of the parts, and use string interpolation to tie it together.
If you're feeling really adventurous, you can have the builder [generate a Roslyn AST](https://gist.github.com/cmendible/9b8c7d7598f1ab0bc7ab5d24b2622622), which would give you a very flexible output generation.

Good job on the library!